Update VHO journey
