+++ VideoWeb/VideoWeb.AcceptanceTests/Journeys/VhoJourney.cs
@@ -11,6 +11,7 @@ namespace VideoWeb.AcceptanceTests.Journeys
             return new List<Page>()
             {
                 Page.Login,
+                Page.VhoVenueList,
                 Page.VhoHearingList
             };
         }
